move repl to its own folder
