+++ package.json
@@ -18,7 +18,7 @@
     "scripts": {
         "build": "tsc",
         "test": "tsx --test tests/**/*.test.ts",
-        "start": "tsx src/index.ts"
+        "start": "tsx src/repl/index.ts"
     },
     "devDependencies": {
         "@types/node": "^24.3.0",
