+++ src/repl/README.md
@@ -1,0 +1,1 @@
+This is a simple Read-Eval-Print-Loop to test the interpreter.
